dis.c - fix 32-bit compile (blob_size, size_t, "%ld")
The disassembler used a size_t for the variable that holds the blob
size, even though it can't ever be bigger than 2^32.

There's also a printf statement that prints the blob size with a %ld
specifier (for a long int), which happens to work fine for size_t on
a 64-bit Linux platform.

It doesn't work on a 32-bit platform where size_t is 32-bit.

The current fix is not the right one.  This and many other incorrect
printf format strings for uint32_t et al will be fixed in ReVAX v0.2.

+++ src/dis.c
@@ -96,7 +96,7 @@ STATIC struct dis_ret op_dis_sane(uint8_t b[MAX_OPLEN], uint32_t pc, int width, 
 
 uint32_t	blob_start;
 uint8_t		blob[1024*1024];
-size_t		blob_size;
+uint32_t	blob_size;
 
 
 /***/
@@ -1166,7 +1166,7 @@ int main(int argc, char *argv[])
 	detect_type(fname);
 	load_blob(fname);
 	blob_start = 0x20040000;
-printf("blob_size: %ld\n", blob_size);
+printf("blob_size: %d\n", blob_size);
 
 /* FIXME "written runtime/ka655x.bin.asm"
          "written runtime/ka655x.bin.lst"
